Memory-leak demo... 3 ways
This is now rebased onto the cereal-1.3.0 update.  I still see leaks
with all of the following:

- examples/test_memory_leak.py test1
- examples/test_memory_leak.py test2
- bin/test_memory_leak

I am finding the issue currently occurs only when caching
G3TimestreamMap objects from frames, not simple G3VectorDouble.

This demo also has an "so3g" module with the leaker code in it, but
this is not in use -- the tests above all default to using
core.G3MemoryLeaker.